requirements.txt
paddleocr==2.7.0.2
paddleocr.egg==info
Pillow==10.0.0
Pillow==10.0.0
PyMySQL==1.1.0
